# specs/001-comidapp-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ComidApp API
  version: 1.0.0
  description: >
    API REST para ComidApp — plataforma web de pedidos y entrega de comida rápida.
    Todos los endpoints protegidos requieren token JWT en el header Authorization.
  contact:
    name: ComidApp Team
servers:
  - url: http://localhost:3001/api/v1
    description: Desarrollo local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EstadoPedido:
      type: string
      enum: [RECIBIDO, EN_PREPARACION, EN_CAMINO, ENTREGADO, CANCELADO]
    MetodoPago:
      type: string
      enum: [TARJETA, EFECTIVO]
    EstadoPago:
      type: string
      enum: [PENDIENTE, APROBADO, RECHAZADO, REVERTIDO]
...
    ErrorResponse:
      type: object
      required: [error, mensaje]
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        mensaje:
          type: string
          example: El campo email es requerido
...
    Local:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
          example: El Monstruo
        direccion:
          type: string
        barrio:
          type: string
        tiempoEstimadoMin:
          type: integer
          example: 30
        abierto:
          type: boolean
          description: Indica si el local está dentro de su horario de atención ahora
    HorarioLocal:
      type: object
      properties:
        diaSemana:
          type: integer
          minimum: 0
          maximum: 6
          description: 0=Domingo, 1=Lunes, ..., 6=Sábado
        horaApertura:
          type: string
          example: "11:00"
        horaCierre:
          type: string
          example: "23:00"

    # ── Productos / Menú ─────────────────────────────────────────────────────
    ItemMenu:
      type: object
      properties:
        id:
          type: integer
          description: ID del item_menu (no del producto)
        productoId:
          type: integer
        nombre:
          type: string
        descripcion:
          type: string
          nullable: true
        precio:
          type: number
          format: double
          example: 2500.00
        disponible:
          type: boolean
        imagenUrl:
          type: string
          nullable: true

    # ── Carrito ──────────────────────────────────────────────────────────────
...
    ItemPedido:
      type: object
      properties:
        productoId:
          type: integer
        nombreProducto:
          type: string
        precioUnitario:
          type: number
          format: double
        cantidad:
          type: integer
        subtotal:
          type: number
          format: double
    Pedido:
      type: object
      properties:
        id:
          type: integer
        clienteId:
          type: integer
        localId:
          type: integer
        localNombre:
          type: string
        repartidorId:
          type: integer
          nullable: true
        estado:
          $ref: '#/components/schemas/EstadoPedido'
        metodoPago:
          $ref: '#/components/schemas/MetodoPago'
        precioTotal:
          type: number
          format: double
        costoEnvio:
          type: number
          format: double
        direccionEntrega:
          type: string
        fechaCreacion:
          type: string
          format: date-time
        fechaActualizacion:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemPedido'
        pago:
          $ref: '#/components/schemas/Pago'

    # ── Pagos ────────────────────────────────────────────────────────────────
    Pago:
      type: object
      properties:
        id:
          type: integer
        estado:
          $ref: '#/components/schemas/EstadoPago'
        metodo:
          $ref: '#/components/schemas/MetodoPago'
        monto:
          type: number
          format: double
        externalPaymentId:
          type: string
          nullable: true
paths:
...
  /locales:
    get:
      tags: [Locales]
      summary: Listar los 5 locales con estado de apertura actual (FR-004, FR-033)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de locales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Local'
  /locales/{localId}/horarios:
    get:
      tags: [Locales]
      summary: Obtener horarios de atención de un local
      security:
        - bearerAuth: []
      parameters:
        - name: localId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Horarios del local
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HorarioLocal'
        '404':
          description: Local no encontrado

  # ══════════════════════════════════════════════════════════
  # MENÚ
  # ══════════════════════════════════════════════════════════
  /menu/{localId}:
    get:
      tags: [Menu]
      summary: Obtener menú activo de un local (FR-004, FR-005, FR-006)
      security:
        - bearerAuth: []
      parameters:
        - name: localId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de productos disponibles del local
          content:
            application/json:
              schema:
                type: object
                properties:
                  local:
                    $ref: '#/components/schemas/Local'
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ItemMenu'
        '404':
          description: Local no encontrado
...
  /pedidos:
    post:
      tags: [Pedidos]
      summary: Confirmar pedido desde el carrito activo (FR-011, FR-012, FR-013, FR-033)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [metodoPago]
              properties:
                metodoPago:
                  $ref: '#/components/schemas/MetodoPago'
                direccionEntrega:
                  type: string
                  description: Opcional; si no se envía, usa la del perfil del usuario
      responses:
        '201':
          description: Pedido creado con estado RECIBIDO
          content:
            application/json:
              schema:
                type: object
                properties:
                  pedido:
                    $ref: '#/components/schemas/Pedido'
                  checkoutUrl:
                    type: string
                    nullable: true
                    description: URL de Mercado Pago si metodoPago=TARJETA; null si EFECTIVO
        '400':
          description: Carrito vacío, productos no disponibles o local cerrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [CARRITO_VACIO, PRODUCTOS_NO_DISPONIBLES, LOCAL_CERRADO]
                  detalle:
                    type: string
                  itemsNoDisponibles:
                    type: array
                    items:
                      type: integer
                    description: IDs de item_menu no disponibles (solo para PRODUCTOS_NO_DISPONIBLES)
        '409':
          description: No hay repartidores disponibles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /pedidos/{pedidoId}:
    get:
      tags: [Pedidos]
      summary: Obtener detalle de un pedido (FR-019, FR-031)
      security:
        - bearerAuth: []
      parameters:
        - name: pedidoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalle del pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
        '403':
          description: El usuario no tiene acceso a este pedido
        '404':
          description: Pedido no encontrado
  /pedidos/{pedidoId}/estado:
    patch:
      tags: [Pedidos]
      summary: Actualizar estado del pedido (FR-017, FR-030) — repartidor o sistema
      security:
        - bearerAuth: []
      parameters:
        - name: pedidoId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nuevoEstado]
              properties:
                nuevoEstado:
                  $ref: '#/components/schemas/EstadoPedido'
      responses:
        '200':
          description: Estado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
        '400':
          description: Transición de estado inválida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: TRANSICION_INVALIDA
                  estadoActual:
                    $ref: '#/components/schemas/EstadoPedido'
                  transicionesValidas:
                    type: array
                    items:
                      $ref: '#/components/schemas/EstadoPedido'
        '403':
          description: El usuario no tiene permiso para actualizar este pedido

  # ══════════════════════════════════════════════════════════
  # PAGOS
  # ══════════════════════════════════════════════════════════
...
  /repartidor/pedidos:
    get:
      tags: [Repartidor]
      summary: Listar pedidos asignados al repartidor autenticado (FR-029)
      security:
        - bearerAuth: []
      parameters:
        - name: estado
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/EstadoPedido'
      responses:
        '200':
          description: Pedidos asignados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pedido'
...